added new movie attributes for smartfolder search (mantis 0000005)

+++ trunk/src/main/java/at/ac/tuwien/e0525580/omov/gui/smartfolder/fields/DateInLastField.java
@@ -42,7 +42,7 @@ public class DateInLastField extends AbstractCriterionField {
 
     @Override
     public Object[] getValues() {
-        int number = this.numberField.getNumber();
+        int number = (int) this.numberField.getNumber();
         int value = number * RangeType.getByString((String) this.comboRangeType.getSelectedItem()).getDayAmount();
         return new Integer[] { value };
     }
